🐛 fix(_c_interpolation_catmull_rom.c): rename variables and fix indexing The variables u0 and v0 have been renamed to r_int and c_int, respectively, to improve readability and consistency with the naming conventions. The indexing of the image array has been fixed to use r_neighbor and c_neighbor instead of u and v.
✨ feat(nanopyx): add skip to doctest
The doctest for the MandelbrotBenchmark has been updated to use the SKIP directive instead of ELLIPSIS to avoid running the benchmark during testing.

🔨 refactor(_le_interpolation_catmull_rom_.cl): simplify variable names and remove unnecessary comments
This commit simplifies the variable names in the _c_interpolate function to improve readability. The comments that were removed were unnecessary as the code is self-explanatory.

🔨 refactor(_le_interpolation_nearest_neighbor_.cl): remove unnecessary line breaks and spaces
This commit removes unnecessary line breaks and spaces in the _le_interpolation_nearest_neighbor_.cl file. The code is now more concise and easier to read.