🔀 chore(_le_mandelbrot_benchmark.pyx): refactor MandelbrotBenchmark class This commit refactors the MandelbrotBenchmark class in `_le_mandelbrot_benchmark.pyx`. The `_run_unthreaded` method is tagged with `tag-start` and `tag-end` to indicate the start and end of the method. The `_run_threaded`, `_run_threaded_static`, and `_run_threaded_dynamic` methods are created by copying the `_run_unthreaded` method and replacing the method name and the `j in range(size)` loop with `j in prange(size)` loop. This improves the readability of the code and makes it easier to maintain.
🔧 chore(_le_mandelbrot_benchmark.pyx): optimize MandelbrotBenchmark._run_threaded_guided method
The _run_threaded_guided method in the MandelbrotBenchmark class has been optimized to use parallelization with OpenMP. The method now uses the prange function from the cython.parallel module to parallelize the loop over the columns of the image. The schedule parameter has been set to "guided" to improve performance. The changes have been tagged with # tag-copy to indicate that they were copied from the _run_unthreaded method.

🆕 chore(tag2tag.py): add new script to auto-generate code from tags
This commit adds a new script called tag2tag.py which auto-generates code from tags in Python and Cython files. The script searches for all tags in the specified directory and replaces them with the corresponding code. The tags are defined as follows:
# tag-start: tag_name
... code ...
# tag-end

To replace a tag, use the following format:
# tag-copy: tag_name; replace("old", "new")
... code ...
# tag-end

The script is located in the nanopyx/scripts directory and can be run using the command `python tag2tag.py`.